#15 chore: change relatorio-financeiro's name
