+++ src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { RelatorioDeConsumoDiarioComponent } from './pages/relatorio-de-consumo-diario/relatorio-de-consumo-diario.component';
-import { RelatorioDeConsumoMensalComponent } from './pages/relatorio-de-consumo-mensal/relatorio-de-consumo-mensal.component';
-import { RelatorioFinanceiroMensalComponent } from './pages/relatorio-financeiro-mensal/relatorio-financeiro-mensal.component';
+import { RelatorioDeConsumoDiarioComponent } from './pages/relatorio-de-consumo/relatorio-de-consumo-diario/relatorio-de-consumo-diario.component';
+import { RelatorioDeConsumoMensalComponent } from './pages/relatorio-de-consumo/relatorio-de-consumo-mensal/relatorio-de-consumo-mensal.component';
+import { RelatorioFinanceiroComponent } from './pages/relatorio-financeiro/relatorio-financeiro.component';
 import { SobreComponent } from './pages/sobre/sobre.component';
 
 const routes : Routes = [
@@ -12,7 +12,7 @@ const routes : Routes = [
     {path: 'diario', component: RelatorioDeConsumoDiarioComponent},
     {path: 'mensal', component: RelatorioDeConsumoMensalComponent}
   ]},
-  {path: 'relatorio-financeiro', component: RelatorioFinanceiroMensalComponent},
+  {path: 'relatorio-financeiro', component: RelatorioFinanceiroComponent},
   {path: 'sobre', component: SobreComponent}
 ];
