+++ src/app/app.module.ts
@@ -20,9 +20,9 @@ import { MatNativeDateModule } from '@angular/material/core';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 // Components
 import { SideNavComponent } from './components/side-nav/side-nav.component';
-import { RelatorioDeConsumoDiarioComponent } from './pages/relatorio-de-consumo-diario/relatorio-de-consumo-diario.component';
-import { RelatorioDeConsumoMensalComponent } from './pages/relatorio-de-consumo-mensal/relatorio-de-consumo-mensal.component';
-import { RelatorioFinanceiroMensalComponent } from './pages/relatorio-financeiro-mensal/relatorio-financeiro-mensal.component';
+import { RelatorioDeConsumoDiarioComponent } from './pages/relatorio-de-consumo/relatorio-de-consumo-diario/relatorio-de-consumo-diario.component';
+import { RelatorioDeConsumoMensalComponent } from './pages/relatorio-de-consumo/relatorio-de-consumo-mensal/relatorio-de-consumo-mensal.component';
+import { RelatorioFinanceiroComponent } from './pages/relatorio-financeiro/relatorio-financeiro.component';
 import { NoDataComponent } from './components/no-data/no-data.component';
 import { LoaderComponent } from './components/loader/loader.component';
 import { SobreComponent } from './pages/sobre/sobre.component';
@@ -34,7 +34,7 @@ import { ChartComponent } from './components/chart/chart.component';
     SideNavComponent,
     RelatorioDeConsumoDiarioComponent,
     RelatorioDeConsumoMensalComponent,
-    RelatorioFinanceiroMensalComponent,
+    RelatorioFinanceiroComponent,
     NoDataComponent,
     LoaderComponent,
     SobreComponent,
