+++ src/app/pages/relatorio-de-consumo/relatorio-de-consumo-diario/relatorio-de-consumo-diario.component.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/custom-colors';
+@import '../../../../styles/custom-colors';
 
 .blue-line { 
     width: 30px;
