+++ src/app/pages/relatorio-de-consumo/relatorio-de-consumo-diario/relatorio-de-consumo-diario.component.ts
@@ -1,7 +1,7 @@
 import { Component, HostListener, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { DataService } from '../../services/data.service';
-import { DailyConsumptionData } from '../../interfaces/daily-consumption-data';
+import { DataService } from '../../../services/data.service';
+import { DailyConsumptionData } from '../../../interfaces/daily-consumption-data';
 import { DomSanitizer } from '@angular/platform-browser';
 
 @Component({
