+++ src/app/pages/relatorio-financeiro/relatorio-financeiro.component.spec.ts
@@ -1,5 +1,4 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RelatorioDeConsumoMensalComponent } from './relatorio-de-consumo-mensal.component';
 import { HttpClientModule } from '@angular/common/http';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
@@ -12,12 +11,13 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import moment, { Moment } from 'moment';
 import { Viewport } from 'karma-viewport/dist/adapter/viewport';
 import { NoDataComponent } from '../../components/no-data/no-data.component';
+import { RelatorioFinanceiroComponent } from './relatorio-financeiro.component';
 
 declare const viewport: Viewport;
 
-describe('RelatorioDeConsumoMensalComponent', () => {
-  let component: RelatorioDeConsumoMensalComponent;
-  let fixture: ComponentFixture<RelatorioDeConsumoMensalComponent>;
+describe('RelatorioFinanceiroComponent', () => {
+  let component: RelatorioFinanceiroComponent;
+  let fixture: ComponentFixture<RelatorioFinanceiroComponent>;
   let datepicker: jasmine.SpyObj<MatDatepicker<Moment>>;
 
   beforeEach(async () => {
@@ -34,12 +34,12 @@ describe('RelatorioDeConsumoMensalComponent', () => {
         MatProgressSpinnerModule
       ],
       declarations: [
-        RelatorioDeConsumoMensalComponent,
+        RelatorioFinanceiroComponent,
         NoDataComponent
       ]
     }).compileComponents();
 
-    fixture = TestBed.createComponent(RelatorioDeConsumoMensalComponent);
+    fixture = TestBed.createComponent(RelatorioFinanceiroComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
     datepicker = jasmine.createSpyObj<MatDatepicker<Moment>>(['close']);
@@ -74,13 +74,13 @@ describe('RelatorioDeConsumoMensalComponent', () => {
       fail('Expected an error to be thrown.');
     }
     catch (error) {
-      expect((error as Error).message).toBe('One of the following HTMLElements does not exist: monthlyConsumptionTitleContainer, monthlyConsumptionTitleWrapper, monthlyConsumptionDatePicker.');
+      expect((error as Error).message).toBe('One of the following HTMLElements does not exist: financialReportTitleContainer, financialReportTitleWrapper, financialReportDatePicker.');
     }
   });
 
   it('should update sizes for responsiveness when titleContainer width > 600px', () => {
-    const titleWrapper = document.getElementById('monthlyConsumptionTitleWrapper');
-    const datePicker = document.getElementById('monthlyConsumptionDatePicker');
+    const titleWrapper = document.getElementById('financialReportTitleWrapper');
+    const datePicker = document.getElementById('financialReportDatePicker');
 
     viewport.set('desktop');
 
@@ -97,13 +97,13 @@ describe('RelatorioDeConsumoMensalComponent', () => {
   });
 
   it('should update sizes for responsiveness when titleContainer width <= 600px', () => {
-    const titleWrapper = document.getElementById('monthlyConsumptionTitleWrapper');
-    const datePicker = document.getElementById('monthlyConsumptionDatePicker');
+    const titleWrapper = document.getElementById('financialReportTitleWrapper');
+    const datePicker = document.getElementById('financialReportDatePicker');
 
     viewport.set('mobile');
 
     component.updateSizes();
-
+  
     if (datePicker && titleWrapper) {
       expect(datePicker.classList.contains('wide')).toBeTruthy();
       expect(titleWrapper.classList.contains('wide')).toBeTruthy();
